Move jwt file to lib folder
